refactor(equip-hide): modularize equip_hide.cpp, adopt DMK scanner API, optimize hot path
Split 1869-line equip_hide.cpp into 10 focused modules:
- shared_state, aob_resolver, indexed_string_table, player_detection,
  visibility_write, armor_injection, bald_fix, gliding_fix,
  background_threads, input_handler

Replace manual VirtualQuery walks with DMK::Scanner::scan_executable_regions().
Use DMK::Scanner::find_and_resolve_rip_relative() for IndexedStringA global.
Precompute hidden category mask (30 atomic loads -> 2-3 per classification).
Pre-reserve original vis map to avoid gameplay heap allocations.
